+++ app/routes.ts
@@ -9,7 +9,7 @@ import { setupHomeOfficeDetailsController } from './controllers/home-office-deta
 import { setupIdamController } from './controllers/idam';
 import { setupIndexController } from './controllers/index';
 import { setupPersonalDetailsController } from './controllers/personal-details';
-import { setUpReasonsForAppealController } from './controllers/reason-for-appeal';
+import { setupReasonsForAppealController } from './controllers/reason-for-appeal';
 import { setupStartController } from './controllers/startController';
 import { setupTaskListController } from './controllers/task-list';
 import { setupTypeOfAppealController } from './controllers/type-of-appeal';
@@ -36,7 +36,7 @@ const personalDetailsController = setupPersonalDetailsController({ updateAppealS
 const contactDetailsController = setupContactDetailsController(updateAppealService);
 const checkAndSendController = setupCheckAndSendController(updateAppealService);
 const confirmationController = setConfirmationController();
-const reasonsForAppealController = setUpReasonsForAppealController(updateAppealService);
+const reasonsForAppealController = setupReasonsForAppealController(updateAppealService);
 // not protected by idam
 router.use(healthController);
 router.use(startController);
