+++ views/case-building/reasons/decision-appeal.njk
@@ -14,7 +14,7 @@
         })
         }}
     {% endif %}
-    <form type="submit"  action="/reasons/wrong-decision" method="post">
+    <form type="submit"  action="/reasons/decision-appeal" method="post">
     {{ govukTextarea({
         name: "moreDetail",
         id: "moreDetail",
